agents/log_agent/requirements.txt
redis>=4.5.1
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
crewai>=0.28.0
langchain-openai>=0.0.2
# Log-specific dependencies
kubernetes>=26.1.0
python-dateutil>=2.8.2